emb-engine/requirements.txt
flask
pillow
olefile
pyembroidery
python-dotenv
